Update minimizers.py
A new  function minimizer_second to find set of minimizers has  been added. The minimizer_second duplicates the function minimizer, but it uses a different algorithm for calculations and therefore works faster. A decorator timer has also been added to determine the runtime of functions and other minor edits have been made to the type annotation.